TextBox finished, menus soon tm.
TextBox creation was probably the hardest thing I've had to have done thus far in the project. Not only because I wanted the text to dynamically fit to the bounds of the box but also because of loading the script!

Anyways here's the changes made in this commit:
- The README has been updated to reflect this commit's changes as well as added the fact that Java 11+ should be capable of running the game.
- CharInput has been cleaned up of some non-instance related variables.
- Entity no longer statically calls Config.getConfig(), hopefully reducing the strain.
- EntityType ICON_OVERLAY has been added.
- FadeIn and FadeOut now call Config.getConfig() on load rather than statically.
- FixedTexture's image can now be changed.
- The IconOverlay entity has been added.
- RenderOverlay's been changed to switch the given priority int to a byte on construct rather than passing a byte.
- The TextBox entity has been added.
- TextInput now has QoL methods lockAll(boolean) and allLocked() -> boolean.
- Hangman now calls Script.loadScript() on entry of the program.
- Small JavaDoc fix in Config.
- Script and ScriptSection have been added to the json package.
- ApiKeyEntry has been changed to demo the current script and textbox entity and does not work in the way that it originally did but will be reverted soon.
- MainMenu has been updated to reflect it's superclass Menu.
- Slowly building Menu up.
- Slowly building MenuComponent up.
- ScrollableMenuComponent is a thing. :)
- Sounds from the Splash scene have been removed and the skip has been reduced from 12 to 11.
- Apostrophe has been added to CharMap.
- A new category type (for textures) called Icon has been added to the UI package.
- A new category type (for textures) called Portrait has been added to the UI package.
- Various textures have been added and renamed in the Texture enum.
- TexturePreprocessor's stitch(BufferedImage, BufferedImage) -> BufferedImage has been made public and static.
- ApiRequest has minor documentation added.
- The Script file stored in resources has been added to Constants.
- Minor sound debugging in Scene.
- Various asset additions/updates for:
  - script.json
  - 0to9.pmg
  - apostrophe.png
  - atoz.png
  - bck.png
  - blank_icon.png
  - Executioner.png
  - questionmark.png
  - textbox.png

+++ src/main/java/dev/glitchedcoder/hangman/entity/RenderPriority.java
@@ -12,17 +12,17 @@ public final class RenderPriority implements Comparable<RenderPriority> {
     public static final RenderPriority MAX;
 
     static {
-        MIN = new RenderPriority((byte) -128);
-        LOW = new RenderPriority((byte) -64);
-        NORMAL = new RenderPriority((byte) 0);
-        HIGH = new RenderPriority((byte) 64);
-        MAX = new RenderPriority((byte) 127);
+        MIN = new RenderPriority(-128);
+        LOW = new RenderPriority(-64);
+        NORMAL = new RenderPriority(0);
+        HIGH = new RenderPriority(64);
+        MAX = new RenderPriority(127);
     }
 
     private final byte priority;
 
-    public RenderPriority(byte priority) {
-        this.priority = priority;
+    public RenderPriority(int priority) {
+        this.priority = (byte) priority;
     }
 
     public byte getPriority() {
